Move all LinkedGeo functions to linkedGeo.c
