+++ CMakeLists.txt
@@ -84,6 +84,7 @@ check_vla(have_vla)
 set(LIB_SOURCE_FILES
     src/h3lib/include/bbox.h
     src/h3lib/include/polygon.h
+    src/h3lib/include/polygonAlgos.h
     src/h3lib/include/h3Index.h
     src/h3lib/include/h3UniEdge.h
     src/h3lib/include/geoCoord.h
